[JIRA SONIC-714] PTP Command line support for klish.
This submit includes initial command line support for PTP on klish. This
is a work in progress. At this point the implementation bypasses the
object model as there are still issues with the framework. However, this
submit still uses the klish framework, and the actioner script uses python
swssdk instead of the swagger-rest APIs. In addition, this submit also uses
actioner script that should be accurate when the object model framework
issues are fixed.

Sanity Logs - http://10.59.132.240:9009/projects/csg_sonic/sanity/logs/dev-sanity-th2-td3-02/sonic_dev_sanity_191016_1658_1780

Change-Id: I9c51e30bc886f1f77fbf23531d5c198ebf28b717