Fix DBus registration for host modules
Prior to this change, a second host module Python file would cause the
host service to fail to start, since it would be unable to register
against the DBus endpoint. The fix is to add an additional path, e.g.
`/org/SONiC/HostService/ztp`, during the registration step. This change
also adds a fix in hostQuery function to send the request to the correct
endpoint.