Ztp feature dev (#258)
*  working ztp status without host communication

* Add SONiC Host Service package and infrastructure

This change adds infrastructure to create a host-side server that will
respond to specific requests over D-Bus. The intention is to let clients
that run within a container to query the host for the registered
requests. This also includes infrastructure in Translib to request the
host to perform some service.

This is primarily used for implementing actions that must be done in the
host system, such as image management, show techsupport, ZTP, etc.

*  working backend without cli changes

* CLI changes without clean up

* changes to sonic host service scripts for python3 compatibility

*  cli bug fixes

* ZTP with fully working code

*  Fully tested ztp backend

* working changes for all gets

*  working ztp backend without cli adjustments

* Fully working ztp changes with new cli enhancements

*  Code clean up and additional changes from PR comments

*  Adding temporary fix in common app for post/patch cases with no ABNF schema

* temporary fix for set operations with no ABNF schema

* fix annotation errors

* fixing indentation issues