Config-management, Fix CLI command prompt message  (#306)
* Add SONiC Host Service package and infrastructure

This change adds infrastructure to create a host-side server that will
respond to specific requests over D-Bus. The intention is to let clients
that run within a container to query the host for the registered
requests. This also includes infrastructure in Translib to request the
host to perform some service.

This is primarily used for implementing actions that must be done in the
host system, such as image management, show techsupport, ZTP, etc.

* Initial commit for config management.

* Config-mgmt related changes.
Update sonic-config-mgmt.yang. Replace save/load/reload with copy rpc.
Transition actioner to use new CLI client interface.

* Move cfg_mgt_app.go -> transformer/cfg_mgmt_xfmr.go
Remove host_comm.go

* Add prompt message to confirm reload and load operations.
Add description and pattern information to yang attributes.

* Update hostservice module filename to match the MOD_NAME.

* Update host module, fix help string in cli.xml

* Changing unicode to str, as needed for python3, unicode api not supported in python3.

* Fix CLI command prompt message

* Revert sonic_types.xml changes.