[JIRA SONIC-13089] Openconfig Yang - VxLAN feature
Defined the openconfig yang for VxLAN feature. It is based on the IETF NVO working group. It has
been augemented with openconfig-interfaces, openconfig-network-instances and a new iftype has been
added (IF_NVE) in the openconfig-if-types-ext yang file.

Test Case:
sonic-mgmt-framework has been compiled successfully. Verified that the appropriate yangs are present
in the corresponding yang tree.

Change-Id: Ic74e2460ebc68e9b503035d1691d58007075573a