[JIRA SONIC-13089] Openconfig Yang - VxLAN feature
Made changes to the initial version as follows:

Difference of openconfig-yang as compared to IETF NVO Yang
1. Removed the protocol-bgp field from the vni_instance_entry. Since this will always  be used onlt for
static vxlan feature only.

2. For evpn, the network-instance augment will be used. Introduced the source-nve in network-instance
augment (along with the existing vni-id). This will be used for mapping the vlan/vrf
(network-instance) to vni mapping corresponding to source nve interface.

3. Removed the statistics from the vni-instance since it is not relevant.

4. Reorganized the vni container instance to have openconfig convention where each container has the
config/state sub-containers. Created grouping or the common parameters used for both config/state
containers.

5. Introduced the must/when condition at all the places
a. vni-instance - must condition for the source-nve of interface type IF_NVE
b. vxlan-ni-config - must condition for the source-nve of interface type IF_NVE
c. vxlan-interface augment - when condition for the oc-if-type to be IF_NVE

Test Case(s)
sonic-mgmt-framework has been compiled successfully. Verified that the appropriate yangs are present
in the corresponding yang tree.
Compiled with pyang with lint to ensure it is openconfig compliant.

Change-Id: I3d0ac799bc5f394e9d5e37c4e6f349060a5ec34f