Sonic "show tech-support" Add Translib input validation, enhance error… (#311)
* Sonic "show tech-support" Add Tranlib input validation, enhance error reporting

- Provide input parameter validity checking and error reporting within the
"show tech-support" Translib RPC callback function. This is needed because the
flow of control for RPC commands does not utilize the YGOT infrastructure and
associated input validation.

- Provide enhanced error reporting for each of the types of errors defined in
the host script invoked by the "show tech-support" command to perform
the collection of technical support information.

* Removed an obsolete reference in the Yang file to the Linux date/time format

* Show tech-support: Change parse failure error message to standard format

In response to a posted review comment, I am changing an error message
for a CLI parsing failure to conform to the standard format for
logged errors by prefixing the message with "%Error: ".

* Show tech-support: Change all error logging to standard format

This is a follow-on to the commit made a few minutes ago:
Modify all error messages logged during "show tech-support"
handling to conform to the standard "%Error: ..." format.

* Show tech-support: Modify xfmr error logging to standard format

Apply the same standard used in Python file error logging to
the logs posted for errors detected in xfmr Translib
input validation. (Use "%Error: ...")