[JIRA SONIC-13089] Openconfig Yang - VxLAN feature
Changed all the tables to have config/state containers. (Few of them were not incorporated since key
was the only element in the table).
Included the key of all the tables as leafref from the config table. Key is now part of both config
& state containers.

Test case(s):
sonic-mgmt-framework has been compiled successfully. Verified that the appropriate yangs are present
in the corresponding yang tree.
Compiled with pyang with lint to ensure it is openconfig compliant.

Change-Id: I8a470d7f503f4cd1965664a36e65ae77dc4b1a3d