[JIRA SONIC-16206] Add OC Yang support for remote macs
Added an augmentation to the FDB container of network-instance-l2 to support of remote mac.

The peer container has config container empty. Only the state container contains the remote evpn
peer.

Test Case(s)
Compiled somin-mgmt, generated allyangs_tree.html and ensured that the peer container is present in
FDB which contains the remote-ip in the state container.

Change-Id: I587cf84815cfcf3b40bd5848a3688178c30f37cc