Add facility to save and restore configuration bundle
Prior to this change, the `write memory` command in CLI would only save
the contents of the Config DB. This commit adds support so that files
can be copied from anywhere on the host, or within containers to a
bundle file (e.g. FRR configuration files from the BGP container).

The changes are driven by a configuration file that describes sets of
files and/or folders and the source location - container or host. The
Config DB is always saved and restored, regardless of the contents of
the configuration file.