Python REST client for CLI
1) cli_client.py defines a geneic python REST client for CLI actioners.
Swagger generated client sdk is bulky and CLI uses very very small
subset of the APIs it provides. This generic client is lightweight and
is customized for CLI to management REST server communication.

Sample usage (details are in README_cli_client.md):

  api = cli_client.ApiClient()
  path = cli_client.Path('/restconf/data/sonic-port:sonic-port')
  response = api.get(path)
  if response.ok():
      process_data(response.content)
  else:
      print(response.error_message())

2) cli_log.py defines APIs to write log messages from CLI actioners.
Logs are written to syslog. They are also written to screen if env
variable LOGTOSCREEN is set.