Reorganize patches for klish
The current patching infra for klish involved a custom script as well as
a directory tree of the patches to apply, split for each source file.
This, IMO, is not good for maintainability, since the context/purpose
for each change is split across multiple files and it's a little bit
harder to correlate changes across files. In addition, quilt and stg are
already used for patching elsewhere.

Therefore, replace the individual file diffs with two patches (one for a
possible build fix with libxml, and the other being the Python 3 support
and klish enhancements). Some cleanup was done in the latter patch to
reduce the size of the diff, but should be otherwise functionally
equivalent.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>